Added macro record- and play function, added icons and updated setup-gamelist.xml, updated package.mk to implement needed python-evdev module (#1524)
DESCRIPTION:

"Macro Setup" allows the user to first set a trigger button and after that also a button combo. The configuration/macro is then saved in "macro_config.json" in the "/storage/.config/emuelec/scripts" folder.

"Macro Runner" starts a background process that waits for the trigger button to be activated and then plays the combo that is defined in the .json-config file. If the trigger button is being pressed for 3-5 seconds, the process is being killed and the controller can be used normally again.